requirements.txt
fastapi==0.115.0
sqladmin>=0.18.0
uvicorn[standard]==0.30.6
SQLAlchemy[asyncio]==2.0.36
aiosqlite==0.20.0
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic[email]==2.9.2
python-dotenv==1.0.1
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
pytest==8.3.3
jinja2>=3.1
bcrypt==4.0.1
itsdangerous>=2.1
python-multipart>=0.0.9
pygments==2.19.2
pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.27
anyio>=4.3
pytest-cov>=5.0

